requirements.txt
pytest
GitPython
requests
PyYAML
toml
prompt_toolkit
PyYAML
toml
# prompt_toolkit for multiline input